refactor type checker.
